requirements.txt
absl-py==1.4.0
addict==2.4.0
cachetools==5.3.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
chardet==5.1.0
charset-normalizer==3.0.1
cvxpy==1.3.1
cycler==0.11.0
ecos==2.0.12
efficientnet-pytorch==0.7.1
filelock==3.9.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
huggingface-hub==0.12.0
idna==3.4
imageio==2.9.0
importlib-metadata==6.0.0
joblib==1.2.0
kiwisolver==1.4.4
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.3.4
mdurl==0.1.2
mmcv==2.0.0
mmengine==0.7.3
munch==2.5.0
networkx==3.0
numpy==1.19.5
oauthlib==3.2.2
opencv-python==4.7.0.68
osqp==0.6.2.post8
packaging==23.0
pandas==1.3.5
Pillow==8.3.1
pretrainedmodels==0.7.4
projector-installer==1.8.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.15.1
pyparsing==3.0.9
pypng==0.20220715.0
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
qdldl==0.1.7
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.3.5
rsa==4.9
safetensors==0.3.1
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
scs==3.2.3
segmentation-models-pytorch==0.3.2
six==1.16.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2023.2.3
timm==0.6.12
tomli==2.0.1
torch==1.9.1+cu111
torchaudio==0.9.1
torchvision==0.10.1+cu111
tqdm==4.62.3
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
yapf==0.33.0
zipp==3.12.1
